+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/plus_smart_home_tech/interaction_api/feign/PaymentFeign.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import ru.yandex.practicum.plus_smart_home_tech.interaction_api.dto.order.OrderDto;
 import ru.yandex.practicum.plus_smart_home_tech.interaction_api.dto.payment.PaymentResponseDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @FeignClient(name = "payment", path = "/api/v1/payment")
@@ -16,13 +17,13 @@ public interface PaymentFeign {
     PaymentResponseDto addPayment(@Valid @RequestBody OrderDto orderDto) throws FeignException;
 
     @PostMapping("/totalCost")
-    Double getTotalCost(@Valid @RequestBody OrderDto orderDto) throws FeignException;
+    BigDecimal getTotalCost(@Valid @RequestBody OrderDto orderDto) throws FeignException;
 
     @PostMapping("/refund")
     void paymentSuccess(@RequestBody UUID paymentId) throws FeignException;
 
     @PostMapping("/productCost")
-    Double productCost(@Valid @RequestBody OrderDto orderDto) throws FeignException;
+    BigDecimal productCost(@Valid @RequestBody OrderDto orderDto) throws FeignException;
 
     @PostMapping("/failed")
     void paymentFailed(@RequestBody UUID paymentId) throws FeignException;
